+++ v3.x/user/msg_clock/cmdclk.c
@@ -34,7 +34,7 @@ void CommandClock_Put(char command)
 {
 	static RT_TASK *ackn = 0;
 	unsigned int put = 'a';
-	unsigned int msg;
+	unsigned long msg;
 
 	if (ackn != rt_get_adr(nam2num("CLKTSK"))) {
 		ackn = rt_rpc(rt_get_adr(nam2num("CLKTSK")), put, &msg);
@@ -48,7 +48,7 @@ void CommandClock_Get(char *command)
 {
 	static RT_TASK *ackn = 0;
 	unsigned int get = 'b';
-	unsigned int msg;
+	unsigned long msg;
 
 	if (ackn != rt_get_adr(nam2num("CLKTSK"))) {
 		ackn = rt_rpc(rt_get_adr(nam2num("CLKTSK")), get, &msg);
@@ -60,7 +60,7 @@ void CommandClock_Get(char *command)
 void *CommandClock_task(void *args)
 {
 	RT_TASK *mytask;
-	unsigned int command;
+	unsigned long command;
 	char R = 'R';
 	int ackn = 0;
 	RT_TASK *get = (RT_TASK *)0, *put = (RT_TASK *)0, *task;
